backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
openai
requests
aiosqlite
greenlet
httpx
mcp
e2b-code-interpreter
firebase-admin
asyncpg
alembic
